+++ bsp/stm32/libraries/HAL_Drivers/drv_gpio.c
@@ -8,6 +8,7 @@
  * 2018-11-06     balanceTWK        first version
  * 2019-04-23     WillianChan       Fix GPIO serial number disorder
  * 2020-06-16     thread-liu        add STM32MP1
+ * 2020-09-01     thread-liu        add GPIOZ  
  */
 
 #include <board.h>
@@ -204,6 +205,24 @@ static const struct pin_index pins[] =
     __STM32_PIN(173, K, 13),
     __STM32_PIN(174, K, 14),
     __STM32_PIN(175, K, 15),
+#if defined(GPIOZ)
+    __STM32_PIN(176, Z, 0),
+    __STM32_PIN(177, Z, 1),
+    __STM32_PIN(178, Z, 2),
+    __STM32_PIN(179, Z, 3),
+    __STM32_PIN(180, Z, 4),
+    __STM32_PIN(181, Z, 5),
+    __STM32_PIN(182, Z, 6),
+    __STM32_PIN(183, Z, 7),
+    __STM32_PIN(184, Z, 8),
+    __STM32_PIN(185, Z, 9),
+    __STM32_PIN(186, Z, 10),
+    __STM32_PIN(187, Z, 11),
+    __STM32_PIN(188, Z, 12),
+    __STM32_PIN(189, Z, 13),
+    __STM32_PIN(190, Z, 14),
+    __STM32_PIN(191, Z, 15),
+#endif /* defined(GPIOZ) */
 #endif /* defined(GPIOK) */
 #endif /* defined(GPIOJ) */
 #endif /* defined(GPIOI) */
