+++ bsp/stm32/libraries/STM32MPxx_HAL/SConscript
@@ -30,7 +30,7 @@ STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_spi.c
 STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_tim.c
 STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_tim_ex.c
 ''')
-    
+
 if GetDepend(['RT_USING_SERIAL']):
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_uart.c']
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_uart_ex.c']
@@ -41,7 +41,7 @@ if GetDepend(['RT_USING_SERIAL']):
 
 if GetDepend(['RT_USING_USB_HOST']) or GetDepend(['RT_USING_USB_DEVICE']):
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_pccard.c']
-    src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_pcd.c'] 
+    src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_pcd.c']
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_pcd_ex.c']
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_hcd.c']
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_ll_usb.c']
@@ -79,9 +79,6 @@ if GetDepend(['RT_USING_AUDIO']):
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_i2s_ex.c']
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_sai.c']
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_sai_ex.c']
-
-if GetDepend(['RT_USING_MTD_NAND']):
-    src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_hal_nand.c']
 
 if GetDepend(['BSP_USING_FMC']):
     src += ['STM32MP1xx_HAL_Driver/Src/stm32mp1xx_ll_fmc.c']
