+++ bsp/stm32/stm32mp157a-st-discovery/.settings/org.eclipse.core.runtime.prefs
@@ -1,0 +1,3 @@
+content-types/enabled=true
+content-types/org.eclipse.cdt.core.asmSource/file-extensions=s
+eclipse.preferences.version=1
