+++ bsp/stm32/stm32mp157a-st-discovery/.settings/projcfg.ini
@@ -1,0 +1,9 @@
+#RT-Thread Studio Project Configuration
+#Tue Sep 01 16:41:17 CST 2020
+output_project_path=D\:/3_work/GitRepositories/rt-thread/bsp/stm32
+mcu_name=STM32MP157A
+hardware_adapter=J-Link
+mcu_base_nano_proj=true
+selected_rtt_version=latest
+project_base_bsp=true
+cfg_version=v2.0
