+++ bsp/stm32/stm32mp157a-st-discovery/board/ports/drv_lptim.c
@@ -19,6 +19,7 @@
 #define LOG_TAG             "drv.lptimer"
 #include <drv_log.h>
 
+#define LED7_PIN  GET_PIN(H, 7)
 LPTIM_HandleTypeDef hlptim1;
 
 void LPTIM1_IRQHandler(void)
@@ -36,7 +37,7 @@ void HAL_LPTIM_AutoReloadMatchCallback(LPTIM_HandleTypeDef *hlptim)
 {
     if(hlptim->Instance == LPTIM1)
     {
-        rt_kprintf("hello rt-thread!\n");
+        HAL_GPIO_TogglePin(GPIOH, GPIO_PIN_7);
     }
 }
 
@@ -52,14 +53,37 @@ static int lptim_control(uint8_t pre_value)
     return RT_EOK;
 }
 
-/**
- * This function initialize the lptim
- */
-static int lptim_init(void)
+static int lptim_start(void)
 {
+
+  /* ### Start counting in interrupt mode ############################# */
+    if (HAL_LPTIM_Counter_Start_IT(&hlptim1, 32767) != HAL_OK)
+    {
+        LOG_D("lptim1 start Counting Error!\n");
+        return -RT_ERROR;
+    }
+    
+    return RT_EOK;
+}
+
+static int lptim_stop()
+{
+   if (HAL_LPTIM_Counter_Stop_IT(&hlptim1) != HAL_OK)
+   {
+        LOG_D("lptim1 stop Error!\n");
+        return -RT_ERROR;
+   }
+    
+    return RT_EOK;
+}
+
+int lptim_init(void)
+{
+    rt_pin_mode(LED7_PIN, PIN_MODE_OUTPUT);
+        
     hlptim1.Instance = LPTIM1;
     hlptim1.Init.Clock.Source = LPTIM_CLOCKSOURCE_APBCLOCK_LPOSC;
-    hlptim1.Init.Clock.Prescaler = LPTIM_PRESCALER_DIV1;
+    hlptim1.Init.Clock.Prescaler = LPTIM_PRESCALER_DIV8;
     hlptim1.Init.UltraLowPowerClock.Polarity = LPTIM_CLOCKPOLARITY_RISING;
     hlptim1.Init.UltraLowPowerClock.SampleTime = LPTIM_CLOCKSAMPLETIME_DIRECTTRANSITION;
     hlptim1.Init.Trigger.Source = LPTIM_TRIGSOURCE_SOFTWARE;
@@ -73,55 +97,51 @@ static int lptim_init(void)
         LOG_D("LPTIM Init Error!\n");
         return -RT_ERROR;
     }
-  /* ### Start counting in interrupt mode ############################# */
-    if (HAL_LPTIM_Counter_Start_IT(&hlptim1, 5000) != HAL_OK)
-    {
-        LOG_D("LPTIM Start Counting Error!\n");
-        return -RT_ERROR;
-    }
     
     return RT_EOK;
 }
-
-static int lptim_deinit()
-{
-   if (HAL_LPTIM_DeInit(&hlptim1) != HAL_OK)
-   {
-        LOG_D("LPTIM Deinit Error!\n");
-        return -RT_ERROR;
-   }
-    
-    return RT_EOK;
-}
+INIT_DEVICE_EXPORT(lptim_init);
 
 static int lptim_sample(int argc, char *argv[])
 {
     if (argc > 1)
     {
-        if (!strcmp(argv[1], "run"))
+        if (!strcmp(argv[1], "start"))
         { 
-           lptim_init(); 
+           lptim_start();
+           return RT_EOK;
         }
         else if (!strcmp(argv[1], "stop"))
         {
-            lptim_deinit();
+            lptim_stop();
+            return RT_EOK;
         }
         else if (!strcmp(argv[1], "set"))
         {
             if (argc > 2)
             {
                lptim_control(atoi(argv[2]));
-            }     
+               return RT_EOK;
+            }
+            else
+            {
+                goto _exit;
+            }
+        }
+        else
+        {
+            goto _exit;
         }
     }
-    else
+_exit:
     {
         rt_kprintf("Usage:\n");
-        rt_kprintf("lptim_sample run            - open lptim, shell will printf 'hello rt-thread'\n");
-        rt_kprintf("lptim_sample set            - set the lptim prescaler, lptim_sample set [0 - 7]\n");
+        rt_kprintf("lptim_sample start    - start lptim, the LED7 will start blink\n");
+        rt_kprintf("lptim_sample stop     - stop lptim, the LED7 will stop blink\n");
+        rt_kprintf("lptim_sample set      - set the lptim prescaler to change LED7 blink frquency, lptim_sample set [0 - 7]\n");
     }
 
-    return RT_EOK;
+    return -RT_ERROR;
 }
 MSH_CMD_EXPORT(lptim_sample, low power timer sample);
