+++ bsp/stm32/stm32mp157a-st-discovery/board/ports/drv_pmic.h
@@ -1,6 +1,6 @@
 /**
   ******************************************************************************
-  * @file    stm32mp15xx__stpmic1.h
+  * @file    stm32mp15xx_eval_stpmic1.h
   * @author  MCD Application Team
   * @brief   stpmu driver functions used for ST internal validation
   ******************************************************************************
@@ -25,80 +25,306 @@
 extern "C" {
 #endif
 
-/* Includes ------------------------------------------------------------------*/
 #include "stm32mp1xx_hal.h"
 
 /* Exported types ------------------------------------------------------------*/
 typedef enum
 {
-  STPMU1_BUCK1=1,
-  STPMU1_BUCK2,
-  STPMU1_BUCK3,
-  STPMU1_BUCK4,
-  STPMU1_LDO1,
-  STPMU1_LDO2,
-  STPMU1_LDO3,
-  STPMU1_LDO4,
-  STPMU1_LDO5,
-  STPMU1_LDO6,
-  STPMU1_VREFDDR,
+    STPMU1_BUCK1=1,
+    STPMU1_BUCK2,
+    STPMU1_BUCK3,
+    STPMU1_BUCK4,
+    STPMU1_LDO1,
+    STPMU1_LDO2,
+    STPMU1_LDO3,
+    STPMU1_LDO4,
+    STPMU1_LDO5,
+    STPMU1_LDO6,
+    STPMU1_VREFDDR,
 }PMIC_RegulId_TypeDef;
 
 /* IRQ definitions */
 typedef enum {
-
 /* Interrupt Register 1 (0x50 for latch) */
-IT_SWOUT_R,
-IT_SWOUT_F,
-IT_VBUS_OTG_R,
-IT_VBUS_OTG_F,
-IT_WAKEUP_R,
-IT_WAKEUP_F,
-IT_PONKEY_R,
-IT_PONKEY_F,
-
+    IT_SWOUT_R,
+    IT_SWOUT_F,
+    IT_VBUS_OTG_R,
+    IT_VBUS_OTG_F,
+    IT_WAKEUP_R,
+    IT_WAKEUP_F,
+    IT_PONKEY_R,
+    IT_PONKEY_F,
 /* Interrupt Register 2 (0x51 for latch) */
-IT_OVP_BOOST,
-IT_OCP_BOOST,
-IT_OCP_SWOUT,
-IT_OCP_OTG,
-IT_CURLIM_BUCK4,
-IT_CURLIM_BUCK3,
-IT_CURLIM_BUCK2,
-IT_CURLIM_BUCK1,
-
+    IT_OVP_BOOST,
+    IT_OCP_BOOST,
+    IT_OCP_SWOUT,
+    IT_OCP_OTG,
+    IT_CURLIM_BUCK4,
+    IT_CURLIM_BUCK3,
+    IT_CURLIM_BUCK2,
+    IT_CURLIM_BUCK1,
 /* Interrupt Register 3 (0x52 for latch) */
-IT_SHORT_SWOUT,
-IT_SHORT_SWOTG,
-IT_CURLIM_LDO6,
-IT_CURLIM_LDO5,
-IT_CURLIM_LDO4,
-IT_CURLIM_LDO3,
-IT_CURLIM_LDO2,
-IT_CURLIM_LDO1,
-
+    IT_SHORT_SWOUT,
+    IT_SHORT_SWOTG,
+    IT_CURLIM_LDO6,
+    IT_CURLIM_LDO5,
+    IT_CURLIM_LDO4,
+    IT_CURLIM_LDO3,
+    IT_CURLIM_LDO2,
+    IT_CURLIM_LDO1,
 /* Interrupt Register 3 (0x52 for latch) */
-IT_SWIN_R,
-IT_SWIN_F,
-IT_RESERVED_1,
-IT_RESERVED_2,
-IT_VINLOW_R,
-IT_VINLOW_F,
-IT_TWARN_R,
-IT_TWARN_F,
-
-IRQ_NR,
+    IT_SWIN_R,
+    IT_SWIN_F,
+    IT_RESERVED_1,
+    IT_RESERVED_2,
+    IT_VINLOW_R,
+    IT_VINLOW_F,
+    IT_TWARN_R,
+    IT_TWARN_F,
+    IRQ_NR,
 } PMIC_IRQn;
 
-/**
-  * @}
-  */ 
+/* Private typedef -----------------------------------------------------------*/
+typedef struct {
+  PMIC_RegulId_TypeDef id;
+  uint16_t *voltage_table;
+  uint8_t  voltage_table_size;
+  uint8_t  control_reg;
+  uint8_t  low_power_reg;
+  uint8_t  rank ;
+  uint8_t  nvm_info ;
+} regul_struct;
 
-/** @defgroup STM32MP15XX_EVAL_STPMU_Exported_Constants Exported Constants
-  * @{
-  */ 
 
-/* Exported constants --------------------------------------------------------*/
+/* Those define should reflect NVM_USER section
+ * For ES Eval Configuration this is specified as
+ * 	0xF7,
+ 0x92,
+ 0xC0,
+ 0x02,
+ 0xFA,
+ 0x30,
+ 0x00,
+ 0x33,
+ * */
+#define NVM_SECTOR3_REGISTER_0  0xF7
+#define NVM_SECTOR3_REGISTER_1  0x92
+#define NVM_SECTOR3_REGISTER_2  0xC0
+#define NVM_SECTOR3_REGISTER_3  0x02
+#define NVM_SECTOR3_REGISTER_4  0xFA
+#define NVM_SECTOR3_REGISTER_5  0x30
+#define NVM_SECTOR3_REGISTER_6  0x00
+#define NVM_SECTOR3_REGISTER_7  0x33
+
+/* nvm_vinok_hyst: VINOK hysteresis voltage
+ 00: 200mV
+ 01: 300mV
+ 10: 400mV
+ 11: 500mV
+ *
+ * nvm_vinok: VINOK threshold voltage
+ 00:  3.1v
+ 01:  3.3v
+ 10:  3.5v
+ 11:  4.5v
+ Otp_ldo4_forced :
+ 0:  LDO4 ranks following OTP_RANK_LDO4<1:0>
+ if VBUS_OTG or SWOUT is turn ON condition
+ 1:  LDO4 follows normal ranking procedure
+
+ nvm_longkeypress:
+ 0:  Turn OFF on long key press inactive
+ 1:  Turn OFF on long key press active
+
+ nvm_autoturnon:
+ 0:  PMIC does not start automatically on VIN rising
+ 1:  PMIC starts automatically on VIN rising
+
+ nvm_cc_keepoff :
+ 0:  short circuit does not turn OFF PMIC
+ 1:  short circuit turn OFF PMIC and keep it OFF till CC_flag is reset
+
+ *
+ */
+#define OTP_VINOK_HYST  	((NVM_SECTOR3_REGISTER_0 & 0xC0) >> 6) 	// nvm_vinok_hyst
+#define OTP_VINOK		  	((NVM_SECTOR3_REGISTER_0 & 0x30) >> 4) 	// nvm_vinok
+#define OTP_LDO4_FORCED     ((NVM_SECTOR3_REGISTER_0 & 0x08) >> 3) 	// Otp_ldo4_forced
+#define OTP_LONGKEYPRESSED  ((NVM_SECTOR3_REGISTER_0 & 0x04) >> 2)	// nvm_longkeypress
+#define OTP_AUTOTURNON  	((NVM_SECTOR3_REGISTER_0 & 0x02) >> 1)	// nvm_autoturnon
+#define OTP_CC_KEEPOFF  	((NVM_SECTOR3_REGISTER_0 & 0x01))		// nvm_cc_keepoff
+
+/*
+ * nvm_rank_buck4:
+ 00: rank0
+ 01: rank1
+ 10: rank2
+ 11: rank3
+ nvm_rank_buck3:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+ nvm_rank_buck2:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+ nvm_rank_buck1:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+ *
+ */
+#define OTP_RANK_BUCK4  	((NVM_SECTOR3_REGISTER_1 & 0xC0) >> 6) 	// nvm_rank_buck4
+#define OTP_RANK_BUCK3		((NVM_SECTOR3_REGISTER_1 & 0x30) >> 4) 	// nvm_rank_buck3
+#define OTP_RANK_BUCK2     	((NVM_SECTOR3_REGISTER_1 & 0x0C) >> 2) 	// nvm_rank_buck2
+#define OTP_RANK_BUCK1  	((NVM_SECTOR3_REGISTER_1 & 0x03))		// nvm_rank_buck1
+
+
+/*
+ * nvm_rank_ldo4:
+ 00: rank0
+ 01: rank1
+ 10: rank2
+ 11: rank3
+ nvm_rank_ldo3:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+ nvm_rank_ldo2:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+ nvm_rank_ldo1:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+ *
+ */
+#define OTP_RANK_LDO4  		((NVM_SECTOR3_REGISTER_2 & 0xC0) >> 6) 	// nvm_rank_ldo4
+#define OTP_RANK_LDO3		((NVM_SECTOR3_REGISTER_2 & 0x30) >> 4) 	// nvm_rank_ldo3
+#define OTP_RANK_LDO2     	((NVM_SECTOR3_REGISTER_2 & 0x0C) >> 2) 	// nvm_rank_ldo2
+#define OTP_RANK_LDO1  		((NVM_SECTOR3_REGISTER_2 & 0x03))		// nvm_rank_ldo1
+
+/*
+ * nvm_clamp_output_buck: Clamp output value to 1.3V max
+ 0: output_buck4<5:0> not clamped
+ 1: output_buck4<5:0> to b011100(1.3V)
+
+ nvm_bypass_mode_ldo3: LDO3 forced bypass mode
+ 0: LDO3 not in bypass mode
+ 1: LDO3 in bypass mode
+
+ nvm_rank_vrefddr:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+
+ nvm_rank_ldo6:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+
+nvm_rank_ldo5:
+ 00:  rank0
+ 01:  rank1
+ 10:  rank2
+ 11:  rank3
+ *
+ */
+#define OTP_CLAMP_OUTPUT_BUCK4  ((NVM_SECTOR3_REGISTER_3 & 0x80) >> 7) 	// nvm_clamp_output_buck4
+#define OTP_BYPASS_MODE_LDO3  	((NVM_SECTOR3_REGISTER_3 & 0x40) >> 6) 	// nvm_bypass_mode_ldo3
+#define OTP_RANK_VREFDDR		((NVM_SECTOR3_REGISTER_3 & 0x30) >> 4) 	// nvm_rank_vrefddr
+#define OTP_RANK_LDO6     		((NVM_SECTOR3_REGISTER_3 & 0x0C) >> 2) 	// nvm_rank_ldo6
+#define OTP_RANK_LDO5  			((NVM_SECTOR3_REGISTER_3 & 0x03))		// nvm_rank_ldo5
+
+/*
+ * nvm_output_buck4: Buck4 default output selection
+ 00:  1.15V
+ 01:  1.2V
+ 10:  1.8V
+ 11:  3.3V
+ nvm_output_buck3: Buck3 default output selection
+ 00:  1.2V
+ 01:  1.8V
+ 10:  3.0V
+ 11:  3.3V
+ nvm_output_buck2: Buck2 default output selection
+ 00:  1.1V
+ 01:  1.2V
+ 10:  1.35V
+ 11:  1.5V
+ nvm_output_buck1: Buck1 default output selection
+ 00:  1.1V
+ 01:  1.15V
+ 10:  1.2V
+ 11:  1.25V
+ *
+ */
+#define OTP_OUTPUT_BUCK4  		((NVM_SECTOR3_REGISTER_4 & 0xC0) >> 6) 	// nvm_output_buck4
+#define OTP_OUTPUT_BUCK3		((NVM_SECTOR3_REGISTER_4 & 0x30) >> 4) 	// nvm_output_buck3
+#define OTP_OUTPUT_BUCK2     	((NVM_SECTOR3_REGISTER_4 & 0x0C) >> 2) 	// nvm_output_buck2
+#define OTP_OUTPUT_BUCK1  		((NVM_SECTOR3_REGISTER_4 & 0x03))		// nvm_output_buck1
+
+/*
+ * [7]	OTP_SWOFF_BY_BOOST_OVP:
+ 0 -> SWOUT will not turnoff bu boost OVP
+ 1 -> SWOUT will be turnoff by BOOST OVP
+
+ [6]	reserved
+
+ [5:4]	nvm_output_ldo3: LDO3 default output selection
+ 00:  1.8V
+ 01:  2.5V
+ 10:  3.3V
+ 11:  output_buck2<4:0>/2 (VTT termination for DDR3 x32, Analog divider implemented in Analog)
+
+ [3:2]	nvm_output_ldo2: LDO2 default output selection
+ 00: 1.8V
+ 01: 2.5V
+ 10: 2.9V
+ 11: 3.3V
+
+ [1:0]	nvm_output_ldo1: LDO1 default output selection
+ 00: 1.8V
+ 01: 2.5V
+ 10: 2.9V
+ 11: 3.3V
+
+ *
+ */
+#define OTP_SWOFF_BY_BOOST_OVP  ((NVM_SECTOR3_REGISTER_5 & 0x80) >> 7) 	// OTP_SWOFF_BY_BOOST_OVP
+#define OTP_OUTPUT_LDO3			((NVM_SECTOR3_REGISTER_5 & 0x30) >> 4) 	// nvm_output_ldo3
+#define OTP_OUTPUT_LDO2     	((NVM_SECTOR3_REGISTER_5 & 0x0C) >> 2) 	// nvm_output_ldo2
+#define OTP_OUTPUT_LDO1  		((NVM_SECTOR3_REGISTER_5 & 0x03))		// nvm_output_ldo1
+
+/*
+ * 	[7:4]	reserved
+ *
+ [3:2]	nvm_output_ldo6: LDO6 default output selection
+ 00:  1.0V
+ 01:  1.2V
+ 10:  1.8V
+ 11:  3.3V
+
+ [1:0]	nvm_output_ldo5: LDO5 default output selection
+ 00:  1.8V
+ 01:  2.5V
+ 10:  2.9V
+ 11 :  3.3V
+ *
+ */
+
+#define OTP_OUTPUT_LDO6     	((NVM_SECTOR3_REGISTER_6 & 0x0C) >> 2) 	// nvm_output_ldo6
+#define OTP_OUTPUT_LDO5  		((NVM_SECTOR3_REGISTER_6 & 0x03))		// nvm_output_ldo5
+
+#define ARRAY_SIZE(x) (sizeof(x) / sizeof((x)[0]))
 #define BIT(_x)  (1<<(_x))
 #define STM32_PMIC_NUM_IRQ_REGS       4
 
@@ -280,33 +506,7 @@ IRQ_NR,
 
 #define PMIC_VERSION_ID             0x10
 
-#define NVM_SECTOR3_REGISTER_7      0x33
-//#define STPMU1_I2C_ADDRESS          ((NVM_SECTOR3_REGISTER_7 & 0x7F) << 1 )
-
-/**
-  * @}
-  */ 
-  
-/** @defgroup STM32MP15XX_EVAL_STPMU_Exported_Functions Exported Functions
-  * @{
-  */
-
-/* Exported functions --------------------------------------------------------*/
-uint8_t STPMU1_Register_Read(uint8_t register_id);
-void STPMU1_Register_Write(uint8_t register_id, uint8_t value);
-void STPMU1_Register_Update(uint8_t register_id, uint8_t value, uint8_t mask);
-void STPMU1_Enable_Interrupt(PMIC_IRQn IRQn);
-void STPMU1_Disable_Interrupt(PMIC_IRQn IRQn);
-void STPMU1_Regulator_Enable(PMIC_RegulId_TypeDef id);
-void STPMU1_Regulator_Disable(PMIC_RegulId_TypeDef id);
-uint8_t STPMU1_Is_Regulator_Enabled(PMIC_RegulId_TypeDef id);
-void STPMU1_Regulator_Voltage_Set(PMIC_RegulId_TypeDef id,uint16_t milivolts);
-uint32_t BSP_PMIC_Init(void);
-uint32_t BSP_PMIC_DeInit(void);
-uint32_t BSP_PMIC_Is_Device_Ready(void);
-uint32_t BSP_PMIC_InitRegulators (void);
-__weak void BSP_PMIC_INTn_Callback(PMIC_IRQn IRQn);
-
+#define STPMU1_I2C_ADDRESS          (NVM_SECTOR3_REGISTER_7 & 0x7F)
 
 #ifdef __cplusplus
 }
